Option to upload Octavia amphora img in RAW format
When deploying an overcloud with Octavia, the amphora image is uploaded
to Glance as QCOW2. When using the RBD backend for Glance and Nova, this
causes the image to be converted to RAW on each deployment. With slower
storage, this leads to load balancer creation to fail with a compute
timeout exception from Octavia.

This patch adds an option to allow amphora image to be automatically
uploaded to Glance in raw format.

Partial-Bug: #1778303
Change-Id: I698ed96881756bb547f5338b147b0c795103acb5