Remove the unhandled_error task from baremetal.yaml
With the recent changing in workflow messaging, it is now possible that
we will get stuck in a loop of unhandled errors. Removing the unhandled_error
task will give us the sanest logs on a failure scenario.

This happens because:
1. The workflow sends an error message, which causes the messaging
   workflow to error. It ends in an error status so that the parent
   workflow doesn't need to fail itself when sending errors.
2. The unhandled_error task is triggered because the message sending
   workflow fails. GOTO 1.

The task-defult error handler was added to these workflows due to the
complexity. However, searching launchpad doesn't find any cases of it being
reported and the generic message doesn't seem particularly uesful. Therefore,
the safest option is to remove it.

Further investigation here might be useful, as a generic error handler
that is reliable could be useful. Work in Mistral to send events to a
Zaqar queue is underway and this will likely be a better solution
(clients can look for workflow failures in the events).

Closes-Bug: 1779097
Depends-On: I5f574923bb3f38b8f71a002e643bf89b52069adc
Change-Id: I388c4b93de473778cc11f580a80426539aeed7e2