git integration for GetOvercloudConfig action
Override the default git commit message so that it's clear that the
config-download commit was run from the Mistral action.

Also, first download the existing config container if one exists. Since
the config-download dir is now managed as a git repo, we want to
preserve the git history, so we need to first reuse the existing
contents of the config container.

Change-Id: I0246918d89dfb564e3d472d4b2f517b5beaf4372