Retry inspect of container images
This change parses the error output of a failed skopeo inspect to
detect if it is a missing image error, then adds tenacity retries for all
other error types.

Failures are being seen in CI due to transient AWS DNS issues, so this change
should make image inspections more resilient.

Change-Id: If5ad384cfc8783c3b10c4cc567e42f06e5c6880b
Closes-Bug: #1779213