Add ability to filter container images to modify
We want to be able to filter out and not attempt modifying any of the
container images that are not originated from the kolla project. This
was previously done by running the container image prepare twice, the
first time with the modify playbook and an exclude list for the images,
the second time without the modify playbook only for the images
previously excluded.  This forced us to maintain a list of the images
to exclude which is not optimal as we add more images. The new approach
relies on the image labels to determine if it can be modified.

Change-Id: Ie7f628bb54c83c78e810320c739afff158234804