Improve validations output for troubleshooting.
Currently, the tripleo validations are being used to
evaluate the status of the environment before/after
certain steps (post-deploy, pre-upgrade, post-upgrade)
However, when obtaining the workflow execution output,
if some of the validations failed there is no information
about which one, and the reason behind.

This patch allows identifying the validation being run
in the run_validation action and publishes the results
for the group or list of validations being run in the
run_validations and run_validation_group workflows.

Change-Id: I81ba3a8c287712e0248476e64da0136fc09e26d5
Story: #2002911