ansible: enable pipelining
SSH pipelining is an quick way of speeding up Ansible by executing
several ansible modules without actual file transfer
The boolean flag defaults to off because some systems have a /etc/sudoers
config that disallows it with sudo with the requiretty setting.
Enabling it so we can improve the performances a little.

Change-Id: Iffaa6548e69d18deaf2fb114041a33434044fc4f