Take WSGIScriptAlias into account in docker healthcheck.
Some services running as vhost, as nova_placement make
use of an  alias in which the wsgi service is mapped.
Currently, the nova_placement healthcheck has hardcoded
the alias in the curl [0] request, but if this alias
changes or other service modifies its from / to another,
then the check will start failing again.

[0] - https://github.com/openstack/tripleo-common/blob/master/healthcheck/nova-placement#L6

Change-Id: I5a325be424740e80eafd6b4abd8eb4b3111740aa
Closes-Bug: #1781623