Generate additional roles with defined roles file
NFV cluster deployment should have the parameters closely
associated with hardware and NUMA spec, in which case, the same
role has to be duplicated to associate the hardware groups. This
patch adds the support to generate the roles_data file with
additional roles from the existing defined roles by only changing
few properties.

openstack overcloud roles generate ComputeOvsDpdk:ComputeOvsDpdkDell \
    ComputeOvsDpdk:ComptueOvsDpdkHp -o roles_data.yaml

This command allows to use the defined role (ComputeOvsDpdk) to be
generated as 2 different roles by changing the 'name' alone. This
reduces the copy paste mistakes in generating multiple roles in a
cluster with same services.
Closes-Bug: #1782322

Change-Id: Ifa60eae1ad09b2ceac207114c40c714a6fc67cbc