Set known hosts when --limit is set.
When the option --limit is specified the tasks will
run only on those hosts passed in the option (or all
except the one passed if patterns are used), for that
reason we can't consider that the facts will be
available for all hosts inside the 'overcloud' group.

This patch appends a new line into the known hosts file
with the available 'overcloud' group hosts instead of
creating a new ssh_known_hosts file.

Change-Id: I8c687c53aa4cff18ac15c332a72fb40ed730e69c
Closes-Bug: #1782168