Fix overwriting downloaded config files
When we wrote downloaded config files over previously existing ones,
we sometimes got garbled files with remainders of the previous state,
e.g.:

step_4:
  keystone_refresh:
    action: exec
    command:
    - keystone
    - pkill
    - --signal
    - USR1
    - httpd
    start_order: 1
    user: root
start_order: 1
    user: root

This was because we opened the file for writing but didn't truncate
its size, so this commit adds O_TRUNC to the flags for opening
files. It's very similar to bug 1434187 which we had a long time ago.

Revert "Fix deploy health checks"
Depends-On: Ia2c12d7455564b6297c5f0934812b10fabbdc914

Change-Id: Ib1d3c68ec3c4048ffc7277daf84834288ea50e48
Closes-Bug: #1783866