Create script tripleo-container-image-prepare
The mistral executor image needs to trigger a container image prepare
via an executable during overcloud stack creation, but adding
python-tripleoclient to the executor image causes significant image
bloat.

This change implements a script equivalent to the command "openstack
tripleo container image prepare" with functionality focused only on
the needs of the caller. The script differs to the full command in the
following ways:
- Only one environment file can be specified, so filtering by
  containerized services is not possible (filtering by roles data is
  still done)
- No output files are generated, this script exists only to do actual
  prepare operations
- --dry-run is not supported
- The printed output is just the image parameter values, not a full
  environment file

Change-Id: I2bd53ac64d6ae7a5bb536fa82b1c70a73d883ba8
Blueprint: container-prepare-workflow