Use https for novnc proxy healthcheck if ssl_only is configured
If nova novnc proxy is configured to ssl only, (see LP 178570)
we need to make sure to also use ssl with the healthcheck script.
With this change we verify if ssl_only is configured in nova.conf
and set https as the proto to use for the novnc healthcheck.

With this change we verify if ssl_only is configured in nova.conf
and set https.

Change-Id: Idd96815b774b94ad5a6ffbb2af4d5b78306ddd12
Related-bug: #1785700
(cherry picked from commit e380c3f396b76edc95d09bac2bdc73005fdd73ad)