Ensure all image parameters have a default set
Now that the PrepareContainerImageParameters action is part of the
deploy workflow, environments/containers-default-parameters.yaml is
overwritten with a dry-run prepare. The output of this prepare is
filtered to *only* those images required for deployment.

However heat requires all image parameters to be populated, regardless
of whether they're required, so this change will ensure that a value
is set for all image parameters by sharing the logic from the
PrepareContainerImageEnv image (which runs during the plan creation
workflow).

Change-Id: Ia2f013714dafac91075456576ec421fbddd822ae
Closes-Bug: #1787268