Update error string to be unicode
We need to be more explicit for python3 that the string is a unicode
string. Otherwise it fails with:

TypeError: a bytes-like object is required, not 'str'

This was a side effect of the patch to address the stdout parsing under
python3. See Id0060a3abbcda8edb6124eb096cb824aaea48396

Change-Id: Ib6dd969dd24c0624028e09fac5abea29f7efa870
Related-Blueprint: python3-support