Allow networks in roles_data to be dict or list
With the changes to make all networks for the overcloud
neutron routed networks it makes sense to change the
schema of 'networks' in roles_data to allow the user to
specify both the networks and the subnet of each network
for a role.

Ref change: I85676ae961d34bf6a986fa813c40380d50f09711

This changes the role yaml validation to allow 'networks'
to be either a list or a dict.

Related: blueprint tripleo-routed-networks-templates
Change-Id: I773a38fd903fe287132151a4d178326a46890969