Adjust tenacity mock
Recent changes in tenacity have broken how we mock the wait call for the
test. This change mocks out the inspect.getargspec call that is used in
tenacity to fix our unit tests for now. I think it might be beneficial
to not actually test the retry functionality in the tripleo-common unit
tests as it seems to be more of a test for tenacity than tripleo-common.

Change-Id: Ia7e9aef0ee8413fa5b09af93f7f89cb34f3f761d
Closes-Bug: #1788802