Fixes OVN docker image prep with Neutron
Currently if one passes the OVN env file to generate the SDN docker
images, neutron server will not be generated because OVN service is
missing in the service list. This patch addresses that and brings it
inline with the behavior of other SDNs (like ODL) where the Neutron
Server image is specific to that SDN and requires generating.

Change-Id: Ic08ff58b10d4fa7116163be1f7fce57879cee8c5
Signed-off-by: Tim Rozet <trozet@redhat.com>