Handle sparse resource registry in build_service_filter
During image prepare, the intent of build_service_filter is to filter
out services which are *known* to be not containerized, either because
they are puppet services, or have been mapped to 'OS::Heat::None'.

However if a partial resource registry is passed in, then this was
also filtering out the services with missing resource_registry
mappings. Since we're now assuming services are containerized by
default, this behaviour was cause many of the issues with missing or
no image parameters being generated.

Closes-Bug: #1727347
Change-Id: Ie65038e3ed73a77aeec9178231ab776ed26a2d84