Don't fail tripleo-bootstrap on package installs
This changes the package install task to not fail if there is a problem
installing the bootstrap packages. This is inline with existing
expectations that we don't require a working package manager with
enabled repos on initial images.

A followup task is added which will cause an error if any of the
required bootstrap packages are not already installed.

This patch also fixes the image yaml files that were not using the
parent package openstack-heat-agents which uncovered this issue.

Change-Id: I5df6193a0d72e8b52a2d95e1827173fc74885d37
Closes-Bug: #1792994
(cherry picked from commit 108942079bacd6ebaa5fa5c7ace99c05faa883a8)