config: ignore missing server_id from the stack
When blacklisting nodes on the overcloud, we don't want to generated
a configuration with these servers.
This patch ignore the server when server_id can't be found in the stack
when generating the configuration of the overcloud.
A warning is shown so the operator knows this server isn't part of the
configuration, probably due to blacklisting.
If getting the server name fails for another reason than a KeyError,
we fail the configuration generation and raise an exception with the
error message.

Change-Id: Ie7660894050e5eca251aaf8c10f0cc7e7d837dfc
Closes-Bug: #1793605