Set prepare neutron_driver from NeutronMechanismDrivers
Since the fix for bug #1727347 landed, this interacts poorly with the
logic which decides whether Opendaylight or OVN are deployed. Since
tripleo-heat-templates driven image prepare doesn't have access to the
actual resource registry, it doesn't know if a service in the roles
data is mapped to a real implementation in the resource registry.

This change switches to determining the neutron driver by using the
value of the NeutronMechanismDrivers parameter.

Change-Id: Ia51a825d11bd9b94d0110f13cdf2a6bbcedf6194
Depends-On: I1fd78fd1eeb071788fc23ba343915ce7d6d14f02
Closes-Bug: #1794103