Fail multiple executions of config-download of the same plan
Add an initial task to the config_download_deploy workflow that queries
for existing executions of the same workflow on the same plan. If any
are found, that means that config-download is already running on the
existing plan, so we should fail the additional one that is trying to
start.

Change-Id: I967ae5528834382b6ef69cc4dac10c893c907fc8