Don't add tarball to config-download dir
GitPyton's repo.index.add defaults to force adding all ignored file
patterns from .gitignore. This was causing the generated
overcloud-config.tar.gz to also be included in the git repo. This caused
the config-download directory to recursivley grow larger in size as each
commit contained a tarball of the whole repo.

Using repo.git.add instead allows the .gitignore to be honored so that
the tarball isn't included in any commits.

Change-Id: I5c7b4fc448d957e30c0ea386795c9b698e9febed
Closes-Bug: #1795949