Change ceph-ansible fetch directory format on upgrade from Q to R
If the name of the tarball containing the ceph-ansible fetch directory
matches the format found in Queens, then rename it to match the format
expected in Rocky. I.e if temporary_dir-20180915-191756.tar.gz is found
then rename it to temporary_dir.tar.gz. This workflow is idempotent and
should be run before upgrading from Queens to Rocky.

Change-Id: I048a5697fed269a6a27e82aca884cb9076fa0785
Fixes-Bug: #1795513