[Pike Only] Do not use namespace var name in docker img template.
The namespace variable name is reserved for a jinja2 built-in
class [0]. So, if no parameter namespace is being passed when the
template is being rendered, we'll end up referencing the
jinja built-in class and the container docker images will be
written as <class 'jinja2.utils.Namespace'>/centos-binary-manila-api.
This patch changes the variable name into docker_namespace.

*It's needed to change the parameter name used in python-tripleoclient
https://review.openstack.org/608648

[0] - http://jinja.pocoo.org/docs/2.10/templates/#namespace
Related-Bug: #1796664

Change-Id: Ia212ad94463541e98d41211947a52e5ddd6a6686