Add AllNodesConfig to config-download group vars
Since I7b865bf82520006eef3ac2f36df34b1f3c34e642 added an AllNodesConfig
output to the overcloud, we can consume that in the group_vars
and add it to the main overcloud group which means the data will be
available to all overcloud nodes.

This may be helpful to solve issues resolving per-service bootstrap
nodes, and also the data there may be more generally useful in
ansible tasks.

Partial-Bug: #1792613
Depends-On: I7b865bf82520006eef3ac2f36df34b1f3c34e642
Change-Id: Idcee177b21e85cff9e0bf10f4c43c71eff9364ec