Add container images for openshift 3.11
This adds the necessary images and tag they appropriately for the
related OpenShift services they will be used by.

origin and node images are not updated anymore on the dockerhub
starting from 3.11 and are replaced by origin-control-plane and
origin-node. As a result, openshift_base_image var was renamed
openshift_prefix, and DockerOpenShiftBaseImage heat param was renamed
DockerOpenShiftControlPlaneImage.

The coreos images are normally hosted on quay.io but were copied over
to the dockerhub to workaround bug #1799396.

Change-Id: Ibbd5ff9d3597f5add440b92a27a2f2f669f7bdbe