Skip deployments with no uuid
If a deployment is not triggered for whatever stack operation is in
progress, instead of having a uuid for the deployment value, Heat just
substitutes the resource name of TripleOSoftwareDeployment.

This needs to be handled in the config download code so that we don't
try and query the Heat API for a deployment with that name which will
just result in a 404 Not found.

Change-Id: If4575498387e9dee94a9ebd75353a2a8bf719072
Closes-Bug: #1798439
Upstream-Stein: https://review.openstack.org/#/c/611421/1
(cherry picked from commit f1f96a88a507d9e8324310cb001dea2e3404f7c2)