Run NetworkDeployment as async task
This commit adds special handling of the NetworkDeployment such that it
will be run as an async task with ansible. This should prevent any
issues where the network configuration causes the ssh connection to drop
and the ansible task to either be unnecessarily retried or failed.

Also added are three variables that can be used to control the async
behavior:

async_deployment: boolean which will toggle running all deployments in
                  async mode.
async_timeout:    timeout in seconds to wait for async tasks
async_poll:       interval in seconds to check async task status

These variables can only be set if running the config-download process
manually, however a future patch could wire them up to Heat parameters.

Change-Id: If1f35980a98a9015ca65f2c6a3e4db04725f1c10
Closes-Bug: #1792343