Run Mistral workflow to make temporary Swift URLs on upgrade
The create_swift_backup_container_plan workflow was getting run
on overcloud deploy but not upgrade because it was called only
on the task split for the clone_git_repo task. This patch ensures
it is run regardless of if the call to clone_git_repo is necessary.
The create_swift_backup_container_plan workflow is used by Ceph and
Swift so in theory this was a bug for Swift upgrades too.

Also, updates rename_ceph_ansible_fetch_directory.yaml's notify_zaqar
task to send a meaningful message, as previously the workflow failed
because it tried to reference a variable as if it were a task.

Change-Id: Ifab82c9a5b0a206c14ae7976d2c1c7f0e9fa3d9d
Closes-Bug: #1799945
(cherry picked from commit d1619ed9eac7ebbf8d8efae1476e1981d0a980e4)