Move some container deps out of base
The ansible, python-shade, python-kubernetes packages account for a
significant about of space on the base image but are not currently
required for all images. Here are the top space hogs out of the base
container:

selinux-policy-minimum 15529233
python2-kubernetes 18007990
selinux-policy-targeted 19280401
python2-babel 20806893
systemd 24402038
python-libs 24713084
binutils 25149789
ansible 47814651
glibc-common 120325207

This change moves ansible and python-shade to the mistral base
container as we execute ansible via mistral. The openstack-selinux
dependency is moved out of the base container to the openstack-base
container as it makes more sense there.

Change-Id: I7405d8b9ea8bb7f93696635f1947470f37cdeff4