Fix config-download timeout
Relying on mistral task timeout to fail the ansible command is awkward,
because it doesn't kill the running command, and it doesn't give an easy
indication of the error without looking at the status info.
Instead, use the "timeout" command on the command line to invoke
ansible, and check the status code to find out if it reached it.

Change-Id: Iec15a06b85722ff6dc9209f0a5d4db333b71496f
Closes-Bug: #1801902
(cherry picked from commit d22611b19d09e20994694eaadab13c2aa8ad4477)