Add image_source to containers
Previously when containers were added to the upstream kolla project or
added to our container list, the container build process might try and
build them when we execute kolla-build.  This change adds a image_source
parameter that is used to understand when we should and should not
attempt to build the container with kolla witht he 'overcloud container
image build' command.

Change-Id: I82d6e09c856f78c0b29d7324c59b5e00bf5ae04d
Closes-BUg: #1801969