Switch stable scenarios to voting
In master we have removed these from the gate as they are non-voting
because we plan on replacing them with standalone versions. However for
stable branches we will want to leave these as voting since they are
defined in the gate in the branch layout.

Change-Id: I7c5c82eafead89c165dd7c131fff82b722b20cb1
Related-Bug: #1802144
(cherry picked from commit f8972f763cc1b7fc499714593957435c1d539a75)