tripleo-container-tag: pull image before tagging
Adding a container pull before the tagging operation so we are sure that
the image is in the local registry.

We were relying on docker-puppet.py script to perform the pull but now
we do the tag with Ansible, it's safer to ensure that the image is here.

This patch runs a container pull before the container tag command.
It works with both podman & docker since we use container_cli.

Change-Id: Ib9d4ae9c2b415e427deb4b0cfc441febaa210f77
Closes-Bug: #1802511