Fix ironic-pxe healthcheck.
When the /run_command file is not found on the
container using the ironic-pxe image, the health
status displays unhealthy even though it's not
true. The reason is because grep can't find
the file and the error makes the whole healthcheck
fail.
This patch redirects the error output so the command
doesn't fail, but the conditional is properly
evaluated.

Change-Id: Ia6141f5499f34cd10618378f448598f74aca6273
Closes-Bug: #1803153