Add advanced deployment options
In order to be able to specify the ansible python interpreter, we need
to be able to pass in some extra options into the deployment workflow.
This change adds support for ansible_python_interpreter to the inventory
code as well as exposing a deployment_options input to the deployment
workflow.

Change-Id: Ieed77f5fe886ef38f447bfa647c8e01009f93d89
Related-Blueprint: python3-support