Specify python interpreter for undercloud
We always gather facts for the undercloud and which should match the
python interpreter that the inventory code was generated with.  This is
overridable with the --deploy-python-interpreter.

There is a single case where these might be mismatched from the
overcloud if a user is running python2.7 mistral containers on a python3
undercloud and uses config-download to deploy against python2.7
hosts. But I beleive this is not going to be a supported use case so we
should be able to default to the python version used to generate the
inventory file or use a specified value from the command line.

Change-Id: I877cf123d3ef4df045c95f8484e0b5f1ae66a8ba
Related-Blueprint: python3-support