Set DeployIdentifier on update/upgrade/ffwd prepare actions
We have separate actions for deploy vs. upgrade prepare, because on
upgrade prepare we don't want to trigger Ansible. DeployIdentifier was
left out from the upgrade prepare action, which got exposed as a bug
after we fixed --skip-deploy-identifier functionality in
Ibab17dcaeebea65135fca4f40562109c90f36c27.

Over time it would be good to further improve code sharing between
`overcloud deploy` and `overcloud update/upgrade/ffwd prepare`, but
for now we need a backportable fix so we just set DeployIdentifier in
the upgrade prepare action.

Additionally it turned out that we never updated the plan env via the
UpdateStackAction as intended, which is now fixed too.

Change-Id: Iefaff07995ae47dcef58739eaf90c4026fe255b0
Closes-Bug: #1804858