Fix AnsibleAction.run() when argument is not a dict
Fix ad hoc commands executed by AnsibleAction.run() when command is not
passed as a dict() but as a normal string, the same way like ordinary
ansible command works.
Check if argument is a dict() and only if it is, use json.dumps() on it.
And not break commands passed as string.

Change-Id: I8b18658bfbae57f264db007e5e81556b0512ef55
Closes-bug: 1805342
(cherry picked from commit 01fb1895a29a6f114045bef231feb6160b65ba80)