Address python3 string issues with subprocess
This patch updates our Popen calls to enable universal newlines for
calls that we parse or consume the output for. Without
univeral_newlines=True, the output is treated as bytes under python3
which leads to issues later where we are using it as strings.

See https://docs.python.org/3/glossary.html#term-universal-newlines

Change-Id: I1a9edd0ebac6f0539ac5c33056c442cfcfafb7f3
Related-Blueprint: python3-support