Get osa container image from tripleo-common defaults
When using the tripleo-deploy-openshift script standalone, we want it
to default to the same image as what's used when calling the script
from TripleO. Parse the default values from the
container_image_prepare_defaults.yaml file and use them to construct
the URI of the openshift-ansible image instead of hardcoding it.

The script first search for the `container_image_prepare_defaults.yaml`
in the location of the rpm-installed tripleo-common, then relative to
the script in case of a tripleo-common checkout, and finally defaults
to the latest upstream image if the file cannot be found.

Change-Id: Ieb65bfcfb7f87599488344fbe91114e9d0a021fc