Introduce a --plan option to replace --config-download-dir
The config download directory is always equivalent to a directory named
after the plan in the /var/lib/mistral directory. We can simplify the
command invocation and make it more user-friendly by passing the plan
name instead of the full path to the config-download directory.

This deprecates the --config-download-dir option in favor of --plan.

Plan name will default to `openshift`.

Change-Id: If425246a97ff59f21164d8cbdf2312de40697036