Add backwards compatible keystone healthcheck link
https://review.openstack.org/#/c/631310/ moved the keystone healthcheck
file that we symlink but this assumes that a matching tripleo-common
package exists for the template when built. Because the tripleo-common
can be out of date, we need to temporarily check for the new file before
symlinking to fix container builds. This patch should be reverted once
we have a promotion.

Change-Id: I5e4166854ec0a3f87577826bc61b703ee80f227a
Closes-Bug: #1813176