Use the yaml stdout_callback for better prepare logging
Currently when the tripleo-modify-role runs a shell script which
produces a lot of output, its logged once as a no-newline-blob, and
again as one line per array item. Neither output is easy read, which makes
debugging harder.

This change switches to the yaml stdout_callback, and a verbosity
level which only prints the script output once.

Change-Id: I85fb8ca40e03f077147871fcd85536c0572d526e
Closes-Bug: #1813684