Make the image pulling optional for tripleo-container-tag role.
The role makes possible the re-use of code at the time of tagging the
container images. However, the tagging varies whether we are deploying
the services or upgrading. In upgrade tasks we use this role to tag the
new container image (from the version we're upgrading to) into the image
being used by pacemaker. So, the value for the container_image
parameter is an image id, not an image name, which can't be pulled.
This patch adds a new parameter, pull_image, which is set to true by
default and will allow us to skip the image pulling for the upgrade_tasks
case.

Change-Id: I6fc69f122b3f48064ed831cd622a617a02a28e02
Related-Bug: #1814104