Authorize undercloud tripleo-admin user
This adds the public key into the .ssh/authorized_keys file
in /home/tripleo-admin. Failure is ignored because this may be running
on an undercloud which doesn't yet have a configured tripleo-admin
user.

This change also refactors the removal of the generate_playbook task
since the playbook now just invokes a role, there is no maintainence
benefit to justify the extra complexity.

Partial-Bug: #1813832
Change-Id: Id2ee912c456d66ed189fd5fdbaa5c1c3627bdf20
Depends-On: I4c8ee04534636622581eb386c01790d6610e7f58