Handle uncompressed layers on image export
There is currently a bug in buildah which results in it committing
uncompressed layers, which has exposed an issue with the exporter that
assumes all layers in containers-storage are compressed.

This change handles exporting an uncompressed layer, and also fixes an
issue where the calc_digest was sometimes corrupted by adding the data
twice.

Change-Id: I25ccac3672ec64481a20bf68f3ffc61fcda7a389
Blueprint: podman-support