Append qemu-img to Mistral executor package list
The Octavia playbook in tripleo-common and run in Mistral executor fails
because qemu-img is not available. This happens when either the operator
explicitly instructs TripleO to convert the amphora qcow2 file to raw
format or when Ceph is enabled.

This patch appends qemu-img to the package list so that the command is
available in the mistral-executor container.

Closes-Bug: #1816382

Change-Id: I5aba5ba0a800aabe6e662a333e3f0436c1b563a0