Specify Octavia amphora image hw_architecture property in Glance
In the case of a multi-architecture cloud, Nova will try to schedule an
Amphora to a non-x86_64 architecture compute node, which will fail the
load balancer creation in Octavia.

This patch explicitly specifies the Amphora image hw_architecture
property when the image is being uploaded to Glance. By doing so, it
hints the Nova scheduler which compute nodes to take into
account when it schedules an Amphora instance.

The default for hw_architecture is x86_64, which resembles the current
behavior since this is the only image type that we ship.

Closes-Bug: #1818563

Change-Id: Ia7be6503a40e08d0d1f7f4d89132c9e9b5bd6704
(cherry picked from commit 2f1c0b89ee15ad95e23e23995aa3340d8db59966)